+++ shared-helpers/src/formKeys.ts
@@ -1,3 +1,5 @@
+import { Language } from "@bloom-housing/backend-core/types"
+
 export const stateKeys = [
   "",
   "AL",
@@ -257,3 +259,11 @@ export const listingFeatures = [
   "mobility",
   "visual",
 ]
+
+export const applicationLanguageKeys = [Language.en, Language.es, Language.zh, Language.vi]
+
+export enum RoleOption {
+  Administrator = "administrator",
+  Partner = "partner",
+}
+export const roleKeys = Object.values(RoleOption)