+++ sites/public/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.4.1-alpha.18](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@4.4.1-alpha.17...@bloom-housing/public@4.4.1-alpha.18) (2022-06-07)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [4.4.1-alpha.17](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@4.4.1-alpha.16...@bloom-housing/public@4.4.1-alpha.17) (2022-06-07)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [4.4.1-alpha.16](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@4.4.1-alpha.15...@bloom-housing/public@4.4.1-alpha.16) (2022-06-07)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
 ## [4.4.1-alpha.15](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@4.4.1-alpha.14...@bloom-housing/public@4.4.1-alpha.15) (2022-06-06)
 
 **Note:** Version bump only for package @bloom-housing/public